Split variables and variable sources
