+++ internal/resolution/variables/variables_test.go
@@ -1,0 +1,13 @@
+package variables_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestVariableSources(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Variables Suite")
+}
