+++ internal/resolution/variablesources/bundles_and_dependencies.go
@@ -7,43 +7,13 @@ import (
 
 	"github.com/blang/semver/v4"
 	"github.com/operator-framework/deppy/pkg/deppy"
-	"github.com/operator-framework/deppy/pkg/deppy/constraint"
 	"github.com/operator-framework/deppy/pkg/deppy/input"
 
 	olmentity "github.com/operator-framework/operator-controller/internal/resolution/entities"
 	"github.com/operator-framework/operator-controller/internal/resolution/util/predicates"
 	entitysort "github.com/operator-framework/operator-controller/internal/resolution/util/sort"
+	olmvariables "github.com/operator-framework/operator-controller/internal/resolution/variables"
 )
-
-type BundleVariable struct {
-	*input.SimpleVariable
-	bundleEntity *olmentity.BundleEntity
-	dependencies []*olmentity.BundleEntity
-}
-
-func (b *BundleVariable) BundleEntity() *olmentity.BundleEntity {
-	return b.bundleEntity
-}
-
-func (b *BundleVariable) Dependencies() []*olmentity.BundleEntity {
-	return b.dependencies
-}
-
-func NewBundleVariable(bundleEntity *olmentity.BundleEntity, dependencyBundleEntities []*olmentity.BundleEntity) *BundleVariable {
-	dependencyIDs := make([]deppy.Identifier, 0, len(dependencyBundleEntities))
-	for _, bundle := range dependencyBundleEntities {
-		dependencyIDs = append(dependencyIDs, bundle.ID)
-	}
-	var constraints []deppy.Constraint
-	if len(dependencyIDs) > 0 {
-		constraints = append(constraints, constraint.Dependency(dependencyIDs...))
-	}
-	return &BundleVariable{
-		SimpleVariable: input.NewSimpleVariable(bundleEntity.ID, constraints...),
-		bundleEntity:   bundleEntity,
-		dependencies:   dependencyBundleEntities,
-	}
-}
 
 var _ input.VariableSource = &BundlesAndDepsVariableSource{}
 
@@ -73,7 +43,7 @@ func (b *BundlesAndDepsVariableSource) GetVariables(ctx context.Context, entityS
 	var bundleEntityQueue []*olmentity.BundleEntity
 	for _, variable := range variables {
 		switch v := variable.(type) {
-		case *RequiredPackageVariable:
+		case *olmvariables.RequiredPackageVariable:
 			bundleEntityQueue = append(bundleEntityQueue, v.BundleEntities()...)
 		}
 	}
@@ -101,7 +71,7 @@ func (b *BundlesAndDepsVariableSource) GetVariables(ctx context.Context, entityS
 		bundleEntityQueue = append(bundleEntityQueue, dependencyEntityBundles...)
 
 		// create variable
-		variables = append(variables, NewBundleVariable(head, dependencyEntityBundles))
+		variables = append(variables, olmvariables.NewBundleVariable(head, dependencyEntityBundles))
 	}
 
 	return variables, nil
