+++ internal/resolution/variablesources/crd_constraints.go
@@ -5,26 +5,11 @@ import (
 	"fmt"
 
 	"github.com/operator-framework/deppy/pkg/deppy"
-	"github.com/operator-framework/deppy/pkg/deppy/constraint"
 	"github.com/operator-framework/deppy/pkg/deppy/input"
 
 	olmentity "github.com/operator-framework/operator-controller/internal/resolution/entities"
+	olmvariables "github.com/operator-framework/operator-controller/internal/resolution/variables"
 )
-
-type BundleUniquenessVariable struct {
-	*input.SimpleVariable
-}
-
-// NewBundleUniquenessVariable creates a new variable that instructs the resolver to choose at most a single bundle
-// from the input 'atMostID'. Examples:
-// 1. restrict the solution to at most a single bundle per package
-// 2. restrict the solution to at most a single bundler per provided gvk
-// this guarantees that no two operators provide the same gvk and no two version of the same operator are running at the same time
-func NewBundleUniquenessVariable(id deppy.Identifier, atMostIDs ...deppy.Identifier) *BundleUniquenessVariable {
-	return &BundleUniquenessVariable{
-		SimpleVariable: input.NewSimpleVariable(id, constraint.AtMost(1, atMostIDs...)),
-	}
-}
 
 var _ input.VariableSource = &CRDUniquenessConstraintsVariableSource{}
 
@@ -62,7 +47,7 @@ func (g *CRDUniquenessConstraintsVariableSource) GetVariables(ctx context.Contex
 	gvkToBundleMap := map[string]map[deppy.Identifier]struct{}{}
 	for _, variable := range variables {
 		switch v := variable.(type) {
-		case *BundleVariable:
+		case *olmvariables.BundleVariable:
 			bundleEntities := []*olmentity.BundleEntity{v.BundleEntity()}
 			bundleEntities = append(bundleEntities, v.Dependencies()...)
 			for _, bundleEntity := range bundleEntities {
@@ -100,7 +85,7 @@ func (g *CRDUniquenessConstraintsVariableSource) GetVariables(ctx context.Contex
 			bundleIDs = append(bundleIDs, bundleID)
 		}
 		varID := deppy.IdentifierFromString(fmt.Sprintf("%s package uniqueness", packageName))
-		variables = append(variables, NewBundleUniquenessVariable(varID, bundleIDs...))
+		variables = append(variables, olmvariables.NewBundleUniquenessVariable(varID, bundleIDs...))
 	}
 
 	for gvk, bundleIDMap := range gvkToBundleMap {
@@ -109,7 +94,7 @@ func (g *CRDUniquenessConstraintsVariableSource) GetVariables(ctx context.Contex
 			bundleIDs = append(bundleIDs, bundleID)
 		}
 		varID := deppy.IdentifierFromString(fmt.Sprintf("%s gvk uniqueness", gvk))
-		variables = append(variables, NewBundleUniquenessVariable(varID, bundleIDs...))
+		variables = append(variables, olmvariables.NewBundleUniquenessVariable(varID, bundleIDs...))
 	}
 
 	return variables, nil
