+++ internal/resolution/variablesources/required_package.go
@@ -6,34 +6,13 @@ import (
 
 	"github.com/blang/semver/v4"
 	"github.com/operator-framework/deppy/pkg/deppy"
-	"github.com/operator-framework/deppy/pkg/deppy/constraint"
 	"github.com/operator-framework/deppy/pkg/deppy/input"
 
 	olmentity "github.com/operator-framework/operator-controller/internal/resolution/entities"
 	"github.com/operator-framework/operator-controller/internal/resolution/util/predicates"
 	"github.com/operator-framework/operator-controller/internal/resolution/util/sort"
+	olmvariables "github.com/operator-framework/operator-controller/internal/resolution/variables"
 )
-
-type RequiredPackageVariable struct {
-	*input.SimpleVariable
-	bundleEntities []*olmentity.BundleEntity
-}
-
-func (r *RequiredPackageVariable) BundleEntities() []*olmentity.BundleEntity {
-	return r.bundleEntities
-}
-
-func NewRequiredPackageVariable(packageName string, bundleEntities []*olmentity.BundleEntity) *RequiredPackageVariable {
-	id := deppy.IdentifierFromString(fmt.Sprintf("required package %s", packageName))
-	entityIDs := make([]deppy.Identifier, 0, len(bundleEntities))
-	for _, bundle := range bundleEntities {
-		entityIDs = append(entityIDs, bundle.ID)
-	}
-	return &RequiredPackageVariable{
-		SimpleVariable: input.NewSimpleVariable(id, constraint.Mandatory(), constraint.Dependency(entityIDs...)),
-		bundleEntities: bundleEntities,
-	}
-}
 
 var _ input.VariableSource = &RequiredPackageVariableSource{}
 
@@ -102,7 +81,7 @@ func (r *RequiredPackageVariableSource) GetVariables(ctx context.Context, entity
 		bundleEntities = append(bundleEntities, olmentity.NewBundleEntity(&resultSet[i]))
 	}
 	return []deppy.Variable{
-		NewRequiredPackageVariable(r.packageName, bundleEntities),
+		olmvariables.NewRequiredPackageVariable(r.packageName, bundleEntities),
 	}, nil
 }
