+++ internal/resolution/variablesources/required_package_test.go
@@ -11,48 +11,9 @@ import (
 	"github.com/operator-framework/operator-registry/alpha/property"
 
 	olmentity "github.com/operator-framework/operator-controller/internal/resolution/entities"
+	olmvariables "github.com/operator-framework/operator-controller/internal/resolution/variables"
 	"github.com/operator-framework/operator-controller/internal/resolution/variablesources"
 )
-
-var _ = Describe("RequiredPackageVariable", func() {
-	var (
-		rpv            *variablesources.RequiredPackageVariable
-		packageName    string
-		bundleEntities []*olmentity.BundleEntity
-	)
-
-	BeforeEach(func() {
-		packageName = "test-package"
-		bundleEntities = []*olmentity.BundleEntity{
-			olmentity.NewBundleEntity(input.NewEntity("bundle-1", map[string]string{
-				property.TypePackage: `{"packageName": "test-package", "version": "1.0.0"}`,
-				property.TypeChannel: `{"channelName":"stable","priority":0}`,
-			})),
-			olmentity.NewBundleEntity(input.NewEntity("bundle-2", map[string]string{
-				property.TypePackage: `{"packageName": "test-package", "version": "2.0.0"}`,
-				property.TypeChannel: `{"channelName":"stable","priority":0}`,
-			})),
-			olmentity.NewBundleEntity(input.NewEntity("bundle-3", map[string]string{
-				property.TypePackage: `{"packageName": "test-package", "version": "3.0.0"}`,
-				property.TypeChannel: `{"channelName":"stable","priority":0}`,
-			})),
-		}
-		rpv = variablesources.NewRequiredPackageVariable(packageName, bundleEntities)
-	})
-
-	It("should return the correct package name", func() {
-		Expect(rpv.Identifier()).To(Equal(deppy.IdentifierFromString(fmt.Sprintf("required package %s", packageName))))
-	})
-
-	It("should return the correct bundle entities", func() {
-		Expect(rpv.BundleEntities()).To(Equal(bundleEntities))
-	})
-
-	It("should contain both mandatory and dependency constraints", func() {
-		// TODO: add this test once https://github.com/operator-framework/deppy/pull/85 gets merged
-		//       then we'll be able to inspect constraint types
-	})
-})
 
 var _ = Describe("RequiredPackageVariableSource", func() {
 	var (
@@ -96,7 +57,7 @@ var _ = Describe("RequiredPackageVariableSource", func() {
 		variables, err := rpvs.GetVariables(context.TODO(), mockEntitySource)
 		Expect(err).NotTo(HaveOccurred())
 		Expect(variables).To(HaveLen(1))
-		reqPackageVar, ok := variables[0].(*variablesources.RequiredPackageVariable)
+		reqPackageVar, ok := variables[0].(*olmvariables.RequiredPackageVariable)
 		Expect(ok).To(BeTrue())
 		Expect(reqPackageVar.Identifier()).To(Equal(deppy.IdentifierFromString(fmt.Sprintf("required package %s", packageName))))
 
@@ -125,7 +86,7 @@ var _ = Describe("RequiredPackageVariableSource", func() {
 		variables, err := rpvs.GetVariables(context.TODO(), mockEntitySource)
 		Expect(err).NotTo(HaveOccurred())
 		Expect(variables).To(HaveLen(1))
-		reqPackageVar, ok := variables[0].(*variablesources.RequiredPackageVariable)
+		reqPackageVar, ok := variables[0].(*olmvariables.RequiredPackageVariable)
 		Expect(ok).To(BeTrue())
 		Expect(reqPackageVar.Identifier()).To(Equal(deppy.IdentifierFromString(fmt.Sprintf("required package %s", packageName))))
